combine EmployeeViewModel and ManagerViewModel.kt
