+++ app/src/main/java/com/map08/shiftapp/CompositionLocalProviders.kt
@@ -4,7 +4,6 @@ import androidx.compose.runtime.compositionLocalOf
 import androidx.navigation.NavHostController
 import com.map08.shiftapp.viewmodels.AuthViewModel
 import com.map08.shiftapp.viewmodels.EmployeeViewModel
-import com.map08.shiftapp.viewmodels.ManagerViewModel
 
 val LocalNavController = compositionLocalOf<NavHostController> {
     error("No NavController provided")
@@ -14,7 +13,4 @@ val LocalAuthViewModel = compositionLocalOf<AuthViewModel> {
 }
 val LocalEmployeeViewModel = compositionLocalOf<EmployeeViewModel> {
     error("No EmployeeProfileViewModel provided")
-}
-val LocalManagerViewModel = compositionLocalOf<ManagerViewModel> {
-    error("No ManagerViewModel provided")
 }
